Handled the fail of a service
