+++ packages/workspace/package.json
@@ -32,11 +32,11 @@
     "@capsulajs/capsulajs-configuration-service": "0.1.12",
     "@capsulajs/capsulajs-transport-providers": "0.1.1",
     "@scalecube/scalecube-microservice": "0.1.3-snapshot.1562076916.11",
-    "rxjs": "6.5.2",
     "uuid": "^3.3.2"
   },
   "devDependencies": {
     "@capsulajs/capsulahub-cdn-emulator": "^0.0.2-alpha.2",
-    "@types/uuid": "^3.4.5"
+    "@types/uuid": "^3.4.5",
+    "rxjs": "6.5.2"
   }
 }
