+++ tsconfig.json
@@ -5,8 +5,8 @@
     "lib": ["dom", "es2015"],
     "declaration": true,
     "strict": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
+    "noUnusedLocals": false,
+    "noUnusedParameters": false,
     "moduleResolution": "node",
     "baseUrl": "./src",
     "jsx": "react",
